changes requested from APIDOC-6
