+++ lib/nexmo_developer/app/webpacker/packs/application.js
@@ -48,6 +48,7 @@ import Sidenav from '../javascript/sidenav/index'
 import Topnav from '../javascript/topnav/index'
 import Careers from '../javascript/careers'
 import LocaleSwitcher from '../javascript/locale_switcher'
+import dataLayer from '../javascript/adobe_analytics/index'
 
 $(document).ready(function() {
   Rails.start()
@@ -62,6 +63,10 @@ $(document).ready(function() {
   new Careers
   new LocaleSwitcher
   new Topnav
+
+  // Data layer variable is assigned to window.digitalData
+  // for adobe launch script to pick it up
+  window.digitalData = dataLayer()
 
   if (document.getElementById('jwt-generator-app')) {
     new Vue({
