+++ lib/nexmo_developer/config/webpack/environment.js
@@ -5,6 +5,12 @@ const { VueLoaderPlugin } = require('vue-loader')
 const vue = require('./loaders/vue')
 const ManifestPlugin = require('webpack-manifest-plugin')
 const customConfig = require('./custom')
+const dotenv = require('dotenv')
+dotenv.config({path: __dirname + '/.env'})
+environment.plugins.insert(
+	"Environment",
+	 new webpack.EnvironmentPlugin(process.env)
+  )
 
 const babelLoader = environment.loaders.get('babel')
 babelLoader.exclude = []
